drm/nouveau/pm: detect the fan type and handle them more correctly
There are 4 ways to drive of fans:
- Non controllable fans: Speed is fixed and not controlled by the driver
- Toggle: You can only drive them fully on or fully off
- PWM: Use NVidia's integrated PWM hw implementation
- I2C: Use a dedicated external i2c device to drive the fan

Toggle fans were as commonly found on chipsets up to the nv50 as PWM ones.

Starting from nv50, a hardware implementation of PWM was cast into the silicon
of each chipset. Seems like toggle was put to an end.

NVidia's PWM implementation is pretty simple, just set a divisor, a duty and
you'll get duty/divisor * max_power to the fan. Just what could have expected.

For some reasons, NVidia puts an i2c IC to control both the temperature and
the fan on some high-end cards. Due to the current limitations of the in-kernel

= Driving a GPIO fan =

Toggle fans are driven directly from the GPIO line. The blob toggles the line
at a high frequency to mimic a PWM dedicated circuit.

TODO: current implementation is just a stub
      write an IRQ-based implementation for those fans

= Driving a fan using NVidia's PWM IC =

Nothing new here. Set-up the divisor and the duty in the right regs and you're
done. Nothing interesting here except that nv50+ have two pairs of divisor/duty
regs and the selection of it is done according to the physical GPIO line
referenced by the PWM_FAN tag in the GPIO table.

TODO: Fix the PWM divisor calculation

= Driving an i2c-based IC fan controller =

What we are trying to do is checking if every IC that is known to be used by
NVidia is present on the SMBus at a specified address. The list was extracted
from nvidia-setting's source code.

If one IC is recognized by one driver, then load it and be done with it.

TODO: push the new in-kernel API we need to control the fan and read
the temperature from Nouveau.

Signed-off-by: Martin Peres <martin.peres@labri.fr>